# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.0.120:8000",
    "paths": {
...
        "/api/v1/hoa-don-nhap-kho": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter import invoice based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Filter Import Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update import invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Update Import Invoice",
                "parameters": [
                    {
                        "description": "Update Import Invoice Data",
                        "name": "Update_Import_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_nhap_kho_update"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new import invoice entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Create Import Invoice",
                "parameters": [
                    {
                        "description": "Import Invoice data",
                        "name": "Import_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_nhap_kho_create"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-nhap-kho/lock": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lock import invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Lock Import Invoice",
                "parameters": [
                    {
                        "description": "Lock Import Invoice Data",
                        "name": "Lock_Import_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_nhap_kho_lock"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-nhap-kho/tra-hang": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return Product for import invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Import Return Product",
                "parameters": [
                    {
                        "description": "Return Product Data",
                        "name": "Return_Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Tra_hang_nhap_kho_request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-nhap-kho/tra-no": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pay Debt for import invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import invoice"
                ],
                "summary": "Import Debt Payment",
                "parameters": [
                    {
                        "description": "Debt Payment Data",
                        "name": "Debt_Payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Tra_no_nhap_kho_request"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "requests.Chi_tiet_hoa_don_nhap_kho_create": {
            "type": "object",
            "required": [
                "chiet_khau",
                "ctsp_id",
                "don_vi_tinh",
                "gia_ban",
                "gia_nhap",
                "han_su_dung",
                "hoa_don_id",
                "ke",
                "la_qua_tang",
                "san_pham_id",
                "so_luong",
                "thanh_tien",
                "upc"
            ],
            "properties": {
                "chiet_khau": {
                    "type": "number"
                },
                "ctsp_id": {
                    "type": "integer"
                },
                "don_vi_tinh": {
                    "type": "string"
                },
                "gia_ban": {
                    "type": "number"
                },
                "gia_nhap": {
                    "type": "number"
                },
                "han_su_dung": {
                    "type": "string"
                },
                "hoa_don_id": {
                    "type": "integer"
                },
                "ke": {
                    "type": "string"
                },
                "la_qua_tang": {
                    "type": "boolean"
                },
                "san_pham_id": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "so_luong": {
                    "type": "integer"
                },
                "thanh_tien": {
                    "type": "number"
                },
                "upc": {
                    "type": "string"
                }
            }
        },
...
        "requests.Chi_tiet_san_pham_nhap_tra": {
            "type": "object",
            "required": [
                "cthd_nhap_kho_id",
                "sku",
                "so_luong_tra"
            ],
            "properties": {
                "cthd_nhap_kho_id": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "so_luong_tra": {
                    "type": "integer"
                }
            }
        },
...
        "requests.Hoa_don_nhap_kho_create": {
            "type": "object",
            "required": [
                "ds_san_pham_nhap",
                "kho_id",
                "ngay_nhap",
                "nha_phan_phoi_id"
            ],
            "properties": {
                "con_lai": {
                    "type": "number"
                },
                "ds_san_pham_nhap": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.Chi_tiet_hoa_don_nhap_kho_create"
                    }
                },
                "ghi_chu": {
                    "type": "string"
                },
                "kho_id": {
                    "type": "integer"
                },
                "ma_hoa_don": {
                    "type": "string"
                },
                "ngay_nhap": {
                    "type": "string"
                },
                "nha_phan_phoi_id": {
                    "type": "integer"
                },
                "so_hoa_don": {
                    "type": "integer"
                },
                "tong_tien": {
                    "type": "number"
                },
                "tra_truoc": {
                    "type": "number"
                }
            }
        },
        "requests.Hoa_don_nhap_kho_lock": {
            "type": "object",
            "required": [
                "hoa_don_id"
            ],
            "properties": {
                "hoa_don_id": {
                    "type": "integer"
                },
                "lock_or_open": {
                    "type": "string",
                    "enum": [
                        "lock",
                        "open"
                    ]
                }
            }
        },
        "requests.Hoa_don_nhap_kho_update": {
            "type": "object",
            "required": [
                "ghi_chu",
                "hoa_don_id",
                "ngay_nhap",
                "tra_truoc"
            ],
            "properties": {
                "ghi_chu": {
                    "type": "string"
                },
                "hoa_don_id": {
                    "type": "integer"
                },
                "ngay_nhap": {
                    "type": "string"
                },
                "tra_truoc": {
                    "type": "number"
                }
            }
        },
...
        "requests.Tra_hang_nhap_kho_request": {
            "type": "object",
            "required": [
                "ds_san_pham_tra",
                "hoa_don_id"
            ],
            "properties": {
                "ds_san_pham_tra": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.Chi_tiet_san_pham_nhap_tra"
                    }
                },
                "hoa_don_id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.Tra_no_nhap_kho_request": {
            "type": "object",
            "required": [
                "hoa_don_id",
                "tien_tra"
            ],
            "properties": {
                "hoa_don_id": {
                    "type": "integer"
                },
                "tien_tra": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}